refactor: move github actions schema into /src directory
