+++ .github/actions/generate_manifest/generate_manifest.py
@@ -21,7 +21,7 @@ import os
 import re
 from datetime import datetime, timezone
 
-from schema import Manifest, PackageExtraMetadata, ReleasePackage
+from otaclient_manifest.schema import Manifest, PackageExtraMetadata, ReleasePackage
 
 SCHEMA_VERSION = "1"
