+++ src/otaclient_manifest/schema.py
